Move mnemonic actions from actions.ts
